requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
pytz==2024.1
Werkzeug==3.0.3
gunicorn==21.2.0
yfinance==0.2.36
pandas==2.1.4